Resolve all errors except the EmitCoercionIfNecessary call
